Перенос рендеринга доски из main.js
